Move XInput plugin in OHCI folder
