+++ src/core/HLE/XAPI/OHCI/XInput/DInput.cpp
@@ -1,3 +1,5 @@
+// This is an open source non-commercial project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
 // ******************************************************************
 // *
 // *    .,-:::::    .,::      .::::::::.    .,::      .:
@@ -7,7 +9,7 @@
 // *  `88bo,__,o,    oP"``"Yo,  _88o,,od8P   oP"``"Yo,
 // *    "YUMMMMMP",m"       "Mm,""YUMMMP" ,m"       "Mm,
 // *
-// *   Cxbx->Win32->CxbxKrnl->EmuDInput.h
+// *   src->core->HLE->XAPI->OHCI->XInput->DInput.cpp
 // *
 // *  This file is part of the Cxbx project.
 // *
@@ -31,25 +33,59 @@
 // *  All rights reserved
 // *
 // ******************************************************************
-#ifndef EMUDINPUT_H
-#define EMUDINPUT_H
+#define _XBOXKRNL_DEFEXTRN_
 
-#define DIRECTINPUT_VERSION 0x0800
-#include <dinput.h>
+#define LOG_PREFIX CXBXR_MODULE::DINP
 
-// ******************************************************************
-// * patch: DInputInit
-// ******************************************************************
-extern bool EmuDInputInit();
-
-// ******************************************************************
-// * patch: DInputCleanup
-// ******************************************************************
-extern void EmuDInputCleanup();
+#include "CxbxKrnl/Emu.h"
+#include "CxbxKrnl/EmuXTL.h"
+#include "CxbxKrnl/EmuShared.h"
+#include "Common/Win32/DInputController.h"
+#include "Logging.h"
 
 // ******************************************************************
-// * patch: DInputPoll
+// * Static Variable(s)
 // ******************************************************************
-extern void EmuDInputPoll(PX_XINPUT_STATE Controller);
+static DInputController g_DInputController;
 
-#endif
+// ******************************************************************
+// * XTL::EmuDInputInit
+// ******************************************************************
+bool XTL::EmuDInputInit()
+{
+    g_EmuShared->GetControllerDInputSettings(&g_DInputController.m_settings);
+
+    g_DInputController.ListenBegin(g_hEmuWindow);
+
+    if(g_DInputController.HasError())
+        return false;
+
+    return true;
+}
+
+// ******************************************************************
+// * XTL::EmuDInputCleanup
+// ******************************************************************
+void XTL::EmuDInputCleanup()
+{
+    g_DInputController.ListenEnd();
+}
+
+//emulated dwPacketNumber for DirectInput controller
+DWORD dwPacketNumber_DirectInput = 0;
+
+// ******************************************************************
+// * XTL::EmuDInputPoll
+// ******************************************************************
+void XTL::EmuDInputPoll(XTL::PX_XINPUT_STATE pXboxController)
+{
+    g_DInputController.ListenPoll(pXboxController);
+	//increment of emulated PacketNumber and report back to Controller.
+	dwPacketNumber_DirectInput++;
+	pXboxController->dwPacketNumber = dwPacketNumber_DirectInput;
+
+    if(g_DInputController.HasError())
+        MessageBox(NULL, g_DInputController.GetError().c_str(), "Cxbx-Reloaded [*UNHANDLED!*]", MB_OK);  // TODO: Handle this!
+
+    return;
+}
