+++ src/core/HLE/XAPI/OHCI/XInput/XInput.cpp
@@ -9,7 +9,7 @@
 // *  `88bo,__,o,    oP"``"Yo,  _88o,,od8P   oP"``"Yo,
 // *    "YUMMMMMP",m"       "Mm,""YUMMMP" ,m"       "Mm,
 // *
-// *   Cxbx->Win32->CxbxKrnl->EmuDInput.cpp
+// *   src->core->HLE->XAPI->OHCI->XInput->DInput.cpp
 // *
 // *  This file is part of the Cxbx project.
 // *
