+++ src/core/HLE/XAPI/OHCI/XInput/XInput.h
@@ -7,7 +7,7 @@
 // *  `88bo,__,o,    oP"``"Yo,  _88o,,od8P   oP"``"Yo,
 // *    "YUMMMMMP",m"       "Mm,""YUMMMP" ,m"       "Mm,
 // *
-// *   Cxbx->Win32->CxbxKrnl->EmuXInput.h
+// *   src->core->HLE->XAPI->OHCI->XInput->XInput.h
 // *
 // *  This file is part of the Cxbx project.
 // *
